Rename android package_name
