+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.orbitingapp",
+    package = "ch.republik",
 )
 
 android_resource(
     name = "res",
-    package = "com.orbitingapp",
+    package = "ch.republik",
     res = "src/main/res",
 )
