+++ android/app/src/main/java/ch/republik/MainActivity.java
@@ -1,4 +1,4 @@
-package com.orbitingapp;
+package ch.republik;
 
 import android.os.Bundle;
 import com.facebook.react.ReactActivity;
