+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "orbitingapp",
+  "name": "republik",
   "displayName": "Republik"
 }
